talktalk-frontend/src/backend/requirements.txt
python-dotenv>=1.0.1
google-genai>=1.0.0
